moves templates and config into libs/templating
