+++ openvair/libs/templating/bridge_virtual_network_template.xml.j2
