+++ openvair/libs/templating/templates/bridge_virtual_network_template.xml.j2
