+++ openvair/libs/templating/templates/netplan_bridge_template.yaml.j2
