+++ openvair/libs/templating/templates/port_netplan_template.yaml.j2
